crypto: ccm-aesni: account for alignment in cra_ctxsize
There are alignment requirements for the context structure.
The allocation size should account for that but currently doesn't, which
is probably a reason for this KASAN report:

> BUG: KASAN: slab-out-of-bounds in aesni_ctr_enc_avx_tfm+0x23/0x70
>
> Read of size 4 at addr ffff88811260da10 by task napi/phy0-89/751

> The buggy address is located 4 bytes to the right of
>
>  allocated 524-byte region [ffff88811260d800, ffff88811260da0c)

This commit brings ccm-aesni aead_alg in line with others in
aesni-intel_glue.c by using `.base.cra_ctxsize = CRYPTO_AES_CTX_SIZE`.